The implementations for login and register seems to be working
We can create accounts, and we can login into these accounts, and as soon as we log in, we can are redirected to the  shopping list page, more things need to be added but this is a great Start. Next, I should add a way for the code to be more secure, for example, not being able to access the Shopping list, Should we separate the fact of what regular users and admin users can do? We also need to ask a way for the user to view their own tasks.